+++ test/event_store/recorded_event_test.exs
@@ -1,6 +1,7 @@
 defmodule Commanded.EventStore.RecordedEventTest do
   use ExUnit.Case
 
+  alias Commanded.EventStore.EnrichedMetadata
   alias Commanded.EventStore.RecordedEvent
   alias Commanded.Helpers.EventFactory
 
@@ -41,21 +42,19 @@ defmodule Commanded.EventStore.RecordedEventTest do
           }
         )
 
-      assert enriched_metadata == %{
-               # Event string-keyed metadata
-               "key1" => "value1",
-               "key2" => "value2",
-               # Standard event fields
-               event_id: event_id,
-               event_number: event_number,
-               stream_id: stream_id,
-               stream_version: stream_version,
-               correlation_id: correlation_id,
-               causation_id: causation_id,
-               created_at: created_at,
-               # Additional field
-               application: ExampleApplication
-             }
+      assert %EnrichedMetadata{
+               event_id: ^event_id,
+               event_number: ^event_number,
+               stream_id: ^stream_id,
+               stream_version: ^stream_version,
+               correlation_id: ^correlation_id,
+               causation_id: ^causation_id,
+               created_at: ^created_at,
+               application: ExampleApplication,
+               handler_name: nil,
+               state: nil,
+               metadata: %{"key1" => "value1", "key2" => "value2"}
+             } = enriched_metadata
     end
   end
 end